集成Azure画画，文件结构调整
