+++ FaceMan.SemanticHub/ModelExtensions/AzureOpenAI/Chat/AzureOpenAIResponseWrapper.cs
@@ -9,7 +9,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace FaceMan.SemanticHub.ModelExtensions.AzureOpenAI
+namespace FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat
 {
     public record AzureOpenAIResponseWrapper
     {
