+++ FaceMan.SemanticHub/ModelExtensions/OpenAI/Chat/OpenAIChatCompletionService.cs
@@ -1,6 +1,5 @@
 ﻿using DocumentFormat.OpenXml.EMMA;
-
-using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI;
+using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
 using FaceMan.SemanticHub.ModelExtensions.QianWen;
 using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 
@@ -8,7 +7,7 @@ using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.ChatCompletion;
 using Microsoft.SemanticKernel.Connectors.OpenAI;
 
-namespace FaceMan.SemanticHub.ModelExtensions.OpenAI
+namespace FaceMan.SemanticHub.ModelExtensions.OpenAI.Chat
 {
     public class OpenAIChatCompletionService : IModelExtensionsChatCompletionService
     {
@@ -27,9 +26,9 @@ namespace FaceMan.SemanticHub.ModelExtensions.OpenAI
         public async Task<ChatMessageContent> GetChatMessageContentsAsync(ChatHistory chatHistory, OpenAIPromptExecutionSettings settings = null, Kernel kernel = null, CancellationToken cancellationToken = default)
         {
             var histroyList = new List<ChatMessage>();
-            ChatParameters chatParameters  = new ChatParameters()
+            ChatParameters chatParameters = new ChatParameters()
             {
-                TopP = settings != null? (float)settings.TopP : (float)1.0,
+                TopP = settings != null ? (float)settings.TopP : (float)1.0,
                 MaxTokens = settings != null ? settings.MaxTokens : 512,
                 Temperature = settings != null ? (float)settings.Temperature : (float)1.0,
                 PresencePenalty = settings != null ? (float)settings.PresencePenalty : (float)0.0,
@@ -98,7 +97,7 @@ namespace FaceMan.SemanticHub.ModelExtensions.OpenAI
                 };
                 histroyList.Add(history);
             }
-          
+
             ModelClient client = new(config.ApiKey, ModelType.OpenAI, _url);
             //返回流式聊天消息内容
             await foreach (var item in client.OpenAI.GetStreamingChatMessageContentsAsync(config.ModelId, histroyList, chatParameters, cancellationToken))
@@ -128,7 +127,7 @@ namespace FaceMan.SemanticHub.ModelExtensions.OpenAI
                 };
                 histroyList.Add(history);
             }
-           
+
             ModelClient client = new(config.ApiKey, ModelType.OpenAI, _url);
             //返回流式聊天消息内容
             await foreach (var item in client.OpenAI.GetStreamingChatMessageContentsAsync(config.ModelId, histroyList, chatParameters, cancellationToken))
