+++ FaceMan.SemanticHub/ModelExtensions/QianWen/Chat/QianWenChatCompletionService.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI;
+using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
 using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 
 using Google.Apis.CustomSearchAPI.v1.Data;
@@ -17,7 +17,7 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace FaceMan.SemanticHub.ModelExtensions.QianWen
+namespace FaceMan.SemanticHub.ModelExtensions.QianWen.Chat
 {
     public class QianWenChatCompletionService : IModelExtensionsChatCompletionService
     {
