+++ FaceMan.SemanticHub/ModelExtensions/XunFei/Chat/XunFeiRequestWrapper.cs
