+++ FaceMan.SemanticHub/ModelExtensions/XunFei/Chat/XunFeiResponseWrapper.cs
