+++ FaceMan.SemanticHub/ModelExtensions/XunFei/XunFeiClient.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI;
 using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 using FaceMan.SemanticHub.ModelExtensions.ZhiPu;
 
@@ -18,7 +17,7 @@ using System.Text.Json.Serialization;
 
 using static FaceMan.SemanticHub.ModelExtensions.XunFei.XunFeiChatCompletionService;
 
-using Usage = FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Usage;
+using Usage = FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat.Usage;
 
 namespace FaceMan.SemanticHub.ModelExtensions.XunFei
 {
