+++ FaceMan.SemanticHub/ModelExtensions/ZhiPu/Chat/ZhiPuChatCompletionService.cs
@@ -2,14 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI;
+using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
 using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.ChatCompletion;
 using Microsoft.SemanticKernel.Connectors.OpenAI;
 
-namespace FaceMan.SemanticHub.ModelExtensions.ZhiPu
+namespace FaceMan.SemanticHub.ModelExtensions.ZhiPu.Chat
 {
     public class ZhiPuChatCompletionService : IModelExtensionsChatCompletionService
     {
