+++ FaceMan.SemanticHub/ModelExtensions/ZhiPu/Chat/ZhiPuResponseWrapper.cs
@@ -3,9 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Azure.AI.OpenAI;
-
-using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI;
-
+using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,7 +11,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace FaceMan.SemanticHub.ModelExtensions.ZhiPu
+namespace FaceMan.SemanticHub.ModelExtensions.ZhiPu.Chat
 {
     public record ZhiPuResponseWrapper
     {
